# src/main/java/com/vacaro/benchlab/service/dto/benchmark/CreateRunRequest.java
package com.vacaro.benchlab.service.dto.benchmark;

public record CreateRunRequest(
    Long implementationId,
    Long datasetId,
    Integer timeoutMs,
    Integer memoryMb,
    Double cpuLimit,
    Integer iterations
) {}