# src/main/resources/swagger/api.yml
openapi: '3.0.1'
info:
  title: 'benchLab'
  version: 0.0.1
servers:
  - url: http://localhost:8080/api
paths:
  /algorithms:
    post:
      operationId: createAlgorithm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAlgorithmRequest'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlgorithmResponse'
  /datasets:
    post:
      operationId: createDataset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDatasetRequest'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetResponse'
  /implementations:
    post:
      operationId: createImplementation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateImplementationRequest'
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImplementationResponse'
  /runs:
    post:
      operationId: createRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRunRequest'
      responses:
        '200':
          description: Queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
  /runs/{id}:
    get:
      operationId: getRun
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Run details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunResponse'
  /benchmarks/compare:
    get:
      operationId: compareBenchmarks
      parameters:
        - in: query
          name: algorithmId
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: datasetId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Comparison
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BenchmarkCompareResponse'
  /benchmarks/timeseries:
    get:
      operationId: benchmarkTimeseries
      parameters:
        - in: query
          name: algorithmId
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: language
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Timeseries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BenchmarkTimeseriesResponse'

components:
  schemas:
    CreateAlgorithmRequest:
      type: object
      required: [name, category, version, complexityDeclared]
      properties:
        name: { type: string }
        category: { type: string }
        version: { type: string }
        complexityDeclared: { type: string }
    AlgorithmResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        category: { type: string }
        version: { type: string }
        complexityDeclared: { type: string }
    CreateDatasetRequest:
      type: object
      required: [type, sizeValue, checksum, datasetVersion]
      properties:
        type: { type: string }
        sizeValue: { type: integer, format: int64 }
        seed: { type: integer, format: int64 }
        checksum: { type: string }
        datasetVersion: { type: string }
    DatasetResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        type: { type: string }
        sizeValue: { type: integer, format: int64 }
        seed: { type: integer, format: int64 }
        checksum: { type: string }
        datasetVersion: { type: string }
    CreateImplementationRequest:
      type: object
      required: [algorithmId, language, sourceCode]
      properties:
        algorithmId:
          type: integer
          format: int64
        language:
          type: string
        sourceCode:
          type: string
        compileConfig:
          type: string
        runtimeConfig:
          type: string
    ImplementationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        algorithmId:
          type: integer
          format: int64
        language:
          type: string
        implementationHash:
          type: string
    CreateRunRequest:
      type: object
      required: [implementationId, datasetId]
      properties:
        implementationId:
          type: integer
          format: int64
        datasetId:
          type: integer
          format: int64
        timeoutMs:
          type: integer
        memoryMb:
          type: integer
        cpuLimit:
          type: number
          format: double
        iterations:
          type: integer
    RunMetricResponse:
      type: object
      properties:
        cpuTimeMs: { type: integer, format: int64 }
        wallTimeMs: { type: integer, format: int64 }
        peakMemoryMb: { type: number, format: double }
        exitCode: { type: integer }
        timedOut: { type: boolean }
        compileMs: { type: integer, format: int64 }
    RunArtifactResponse:
      type: object
      properties:
        stdoutTruncated: { type: string }
        stderrTruncated: { type: string }
        outputSizeBytes: { type: integer, format: int64 }
        artifactChecksum: { type: string }
        technicalLogSummary: { type: string }
    RunResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        status: { type: string }
        traceId: { type: string }
        queuedAt: { type: string }
        startedAt: { type: string }
        finishedAt: { type: string }
        failureReason: { type: string }
        metric:
          $ref: '#/components/schemas/RunMetricResponse'
        artifact:
          $ref: '#/components/schemas/RunArtifactResponse'
    BenchmarkCompareRow:
      type: object
      properties:
        language: { type: string }
        avg: { type: number, format: double }
        stddev: { type: number, format: double }
        p50: { type: integer, format: int64 }
        p95: { type: integer, format: int64 }
        validSamples: { type: integer }
    BenchmarkCompareResponse:
      type: object
      properties:
        algorithmId: { type: integer, format: int64 }
        datasetId: { type: integer, format: int64 }
        rows:
          type: array
          items:
            $ref: '#/components/schemas/BenchmarkCompareRow'
    BenchmarkTimeseriesPoint:
      type: object
      properties:
        finishedAt: { type: string }
        wallTimeMs: { type: integer, format: int64 }
    BenchmarkTimeseriesResponse:
      type: object
      properties:
        algorithmId: { type: integer, format: int64 }
        language: { type: string }
        points:
          type: array
          items:
            $ref: '#/components/schemas/BenchmarkTimeseriesPoint'
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - jwt: []